ResultData에 dataName 추가, 게시물 수정 시 권한 체크

+++ src/main/java/com/yjs/exam/demo/controller/UserArticleController.java
@@ -12,6 +12,8 @@ import com.yjs.exam.demo.util.Ut;
 import com.yjs.exam.demo.vo.Article;
 import com.yjs.exam.demo.vo.ResultData;
 
+import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class UserArticleController {
 	@Autowired
@@ -19,69 +21,114 @@ public class UserArticleController {
 
 	@RequestMapping("/user/article/doAdd")
 	@ResponseBody
-	public ResultData doAdd(String title, String body) {
+	public ResultData<Article> doAdd(HttpSession httpSession, String title, String body) {
+		boolean isLogined = false;
+
+		int loginedMemberId = 0;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
+
 		if (Ut.empty(title)) {
 			return ResultData.from("F-1", "title(을)를 작성해주세요.");
 		}
-
 		if (Ut.empty(body)) {
 			return ResultData.from("F-2", "body(을)를 작성해주세요.");
 		}
 
-		ResultData writeArticleRd = articleService.writeArticle(title, body);
+		ResultData<Integer> writeArticleRd = articleService.writeArticle(loginedMemberId, title, body);
 
-		int id = (int)writeArticleRd.getData1();
-		
+		int id = writeArticleRd.getData1();
+
 		Article article = articleService.getArticle(id);
 
-		return ResultData.newData(writeArticleRd,article);
+		return ResultData.newData(writeArticleRd, "article", article);
 	}
 
 	@RequestMapping("/user/article/getArticles")
 	@ResponseBody
-	public ResultData getArticles() {
+	public ResultData<List<Article>> getArticles() {
 		List<Article> articles = articleService.getArticles();
 
-		return ResultData.from("S-1", "게시물 리스트입니다.", articles);
+		return ResultData.from("S-1", "게시물 리스트입니다.", "articles", articles);
 	}
 
 	@RequestMapping("/user/article/getArticle")
 	@ResponseBody
-	public ResultData getArticle(int id) {
+	public ResultData<Article> getArticle(int id) {
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.", id));
 		}
 
-		return ResultData.from("S-1", Ut.f("%d번 게시물 입니다.", id), article);
+		return ResultData.from("S-1", Ut.f("%d번 게시물 입니다.", id), "article", article);
 	}
 
 	@RequestMapping("/user/article/doDelete")
 	@ResponseBody
-	public String doDelete(int id) {
+	public ResultData<Integer> doDelete(HttpSession httpSession, int id) {
+		boolean isLogined = false;
+
+		int loginedMemberId = 0;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
+
 		Article article = articleService.getArticle(id);
 
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-2", "권한이 없습니다.");
+		}
+
 		if (article == null) {
-			return id + "번 게시물이 존재하지 않습니다.";
+			return ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.", id));
 		}
 
 		articleService.deleteArticle(id);
 
-		return id + "번 게시물이 삭제되었습니다.";
+		return ResultData.from("F-1", Ut.f("%d번 게시물이 삭제되었습니다.", id));
 	}
 
 	@RequestMapping("/user/article/doModify")
 	@ResponseBody
-	public String doModify(int id, String title, String body) {
+	public ResultData<Article> doModify(HttpSession httpSession, int id, String title, String body) {
+		boolean isLogined = false;
+
+		int loginedMemberId = 0;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+
+		if (isLogined == false) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요.");
+		}
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
-			return id + "번 게시물이 존재하지 않습니다.";
+			return ResultData.from("F-1", Ut.f("%d번 게시물이 존재하지 않습니다.", id));
 		}
 
-		articleService.modifyArticle(id, title, body);
+		ResultData actorCanModifyRd = articleService.actorCanModify(loginedMemberId, article);
 
-		return id + "번 게시물이 수정되었습니다.";
+		if (actorCanModifyRd.isFail()) {
+			return actorCanModifyRd;
+		}
+
+		return articleService.modifyArticle(id, title, body);
 	}
 }
